+++ Resources.php
@@ -729,18 +729,6 @@ return [
 			],
 		],
 
-	'ext.srf.filtered.list-view' => $formatModule + [
-			'dependencies' => [
-				'ext.srf.filtered',
-			],
-		],
-
-	'ext.srf.filtered.table-view' => $formatModule + [
-			'dependencies' => [
-				'ext.srf.filtered',
-			],
-		],
-
 	'ext.srf.filtered.calendar-view.messages' => $formatModule + $calendarMessages,
 
 	'ext.srf.filtered.calendar-view' => $formatModule +  array(
@@ -748,7 +736,6 @@ return [
 				'filtered/resources/css/ext.srf.filtered.calendar-view.css',
 			),
 			'dependencies' => array(
-				'ext.srf.filtered',
 				'ext.srf.filtered.calendar-view.messages',
 				'ext.jquery.fullcalendar'
 			),
@@ -767,37 +754,41 @@ return [
 			'styles' => [
 				'filtered/resources/css/ext.srf.filtered.map-view.css',
 			],
-			'dependencies' => [
-				'ext.srf.filtered',
-				'ext.srf.filtered.map-view.leaflet',
-			],
 		],
 
 	'ext.srf.filtered.value-filter' => $formatModule + [
-			'styles' => [
-				'filtered/resources/css/ext.srf.filtered.value-filter.css',
-			],
-			'dependencies' => [
-				'ext.srf.filtered',
-			],
+			'styles' => [ 'filtered/resources/css/ext.srf.filtered.value-filter.css' ],
+			'messages' => [
+				'srf-filtered-value-filter-placeholder',
+				'srf-filtered-value-filter-and',
+				'srf-filtered-value-filter-or',
+				],
+			'dependencies' => [ 'ext.srf.filtered.value-filter.select' ],
+		],
+
+	'ext.srf.filtered.value-filter.select' => $formatModule + [
+			'scripts' => [ 'filtered/resources/js/ext.srf.filtered.select.js' ],
+			'styles' => [ 'filtered/resources/css/ext.srf.filtered.select.css' ],
+		],
+
+	'ext.srf.filtered.slider' => $formatModule + [
+			'scripts' => [ 'filtered/resources/js/ext.srf.filtered.slider.js' ],
+			'styles' => [ 'filtered/resources/css/ext.srf.filtered.slider.css' ],
+		],
+
+	'ext.srf.filtered.slider' => $formatModule + [
+			'scripts' => [ 'filtered/resources/js/ext.srf.filtered.slider.js' ],
+			'styles' => [ 'filtered/resources/css/ext.srf.filtered.slider.css' ],
 		],
 
 	'ext.srf.filtered.distance-filter' => $formatModule + [
 			'styles' => [ 'filtered/resources/css/ext.srf.filtered.distance-filter.css' ],
-			'dependencies' => [
-				'ext.srf.filtered',
-				'jquery.ui.slider',
-			],
+			'dependencies' => [ 'ext.srf.filtered.slider' ],
 		],
 
 	'ext.srf.filtered.number-filter' => $formatModule + [
-			'styles' => [
-				'filtered/resources/css/ext.srf.filtered.number-filter.css',
-			],
-			'dependencies' => [
-				'ext.srf.filtered',
-				'jquery.ui.slider',
-			],
+			'styles' => [ 'filtered/resources/css/ext.srf.filtered.number-filter.less' ],
+			'dependencies' => [ 'ext.srf.filtered.slider' ],
 		],
 
 	// Slideshow